Update .gitignore and Propose initial File system
Main.py provides a demonstration of my implementation of the file
system. I also provided file_system_test.py, which tests most of the
functions implemented in file_system.py.

File system is dependent on implementations of Member, Provider, and
Service so I wrote an implementation that basically makes them act as
structs of data.

The tests are dependent on the provided csv files to pass.

Running any of the files will create three diretories: member_reports,
provider_reports, and summary_reports so those are added to the
.gitignore file