Add all reports to File System and Editing Ability
You can now get member reports, provider reports, etf reports, and
manger reports as strings. Not sure if we want to save these anywhere as
files.

Also you can update provider and members.

I added an employee csv file so you can see if an id is a valid employee
id and if they are a manager.